requirements.txt
yt-dlp>=2024.1.0
openai>=1.12.0
google-generativeai>=0.5.0
pydantic>=2.5.0
pyyaml>=6.0
pillow>=10.0.0
mediapipe>=0.10.9
flask>=3.0.0
python-dotenv>=1.0.0
